src/main/java/edu/stanford/irt/laneweb/servlet/binding/VersionDataBinder.java
package edu.stanford.irt.laneweb.servlet.binding;

import java.util.Map;

import javax.servlet.http.HttpServletRequest;

import edu.stanford.irt.laneweb.model.Model;

public class VersionDataBinder implements DataBinder {

    private String version;

    public VersionDataBinder(final String version) {
        this.version = version;
    }

    @Override
    public void bind(final Map<String, Object> model, final HttpServletRequest request) {
        model.put(Model.VERSION, this.version);
    }
}
